+++ R/codedoc_lines.R
@@ -3,13 +3,13 @@
 #' @title `codedoc` Comment Block Lines
 #' @description
 #' Tools to extract comment blocks and return a vector of strings (lines).
-#' @param detect_allowed_keys see `[codedoc::extract_keyed_comment_blocks_]`
-#' @param text_file_paths see `[codedoc::extract_keyed_comment_blocks_]`
+#' @param detect_allowed_keys see `[codedoc::extract_keyed_comment_blocks]`
+#' @param text_file_paths see `[codedoc::extract_keyed_comment_blocks]`
 #' @param extract_arg_list
 #'
 #' - `NULL`: no additional args
-#' - `list`: these args are passed to `[codedoc::extract_keyed_comment_blocks_]`
-#' @param assertion_type see `[dbc::report_to_assertion]`
+#' - `list`: these args are passed to `[codedoc::extract_keyed_comment_blocks]`
+#' @template arg_assertion_type
 #' @name codedoc_lines
 NULL
 
@@ -40,7 +40,7 @@ codedoc_lines <- function(
   extract_arg_list <- as.list(extract_arg_list)
   extract_arg_list[["detect_allowed_keys"]] <- detect_allowed_keys
   extract_arg_list[["text_file_paths"]] <- text_file_paths
-  df <- do.call(codedoc::extract_keyed_comment_blocks_, extract_arg_list,
+  df <- do.call(codedoc::extract_keyed_comment_blocks, extract_arg_list,
                 quote = TRUE)
   unlist(lapply(df[["comment_block"]], function(x) {
     c("", x, "")