+++ man-roxygen/arg_assertion_type.R
@@ -1,0 +1,2 @@
+
+#' @param assertion_type see `[dbc::report_to_assertion]`